+++ OnlySnarf/classes/promotion.py
@@ -1,3 +1,5 @@
+# TODO: finish this later
+
 import re
 from datetime import datetime
 ##